Move arp_common into arp_linux
