+++ collector/arp_linux.go
@@ -11,6 +11,25 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 )
 
+type arpCollector struct {
+	count *prometheus.Desc
+}
+
+func init() {
+	Factories["arp"] = NewArpCollector
+}
+
+// NewArpCollector returns a new Collector exposing ARP stats.
+func NewArpCollector() (Collector, error) {
+	return &arpCollector{
+		count: prometheus.NewDesc(
+			prometheus.BuildFQName(Namespace, "arp",
+				"count"), "ARP entries by device",
+			[]string{"device"}, nil,
+		),
+	}, nil
+}
+
 func getArpEntries() (map[string]uint32, error) {
 	file, err := os.Open(procFilePath("net/arp"))
 	if err != nil {
